requirements.txt
Django==5.2.5
google-ads==28.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-core==2.25.1
requests==2.32.4
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.1.0
pandas>=2.3.0
numpy>=2.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
plotly>=6.0.0
